backend/requirements.txt
fastapi
uvicorn
python-multipart
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-dateutil
python-dotenv
pydantic
pydantic[email]
requests
groq
aiohttp
sqlalchemy
alembic
starlette  # Required for CORS middleware
cryptography  
async-timeout
httpx
feedparser
beautifulsoup4